+++ R/ARMadd.R
@@ -21,7 +21,7 @@
 #' \item{Response Options}{Eight response options. }
 #'     }
 #' @title ICAR Arithmetic Generator
-#' @seealso \code{\link{lisy}}, \code{\link{twoD}}
+#' @seealso \code{\link{lisy}}, \code{\link{spatial2d}}, \code{\link{spatial3d}}
 #' @examples
 #'
 #' arith(numProbs=4, numOps = 4, magnitude = 3, arithOps = "subtract",